requirements.txt
asgiref==3.2.7
Django==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
lxml==4.5.0
Pillow==7.1.1
psycopg2==2.8.4
PyJWT==1.7.1
PyPDF2==1.26.0
python-docx==0.8.10
pytz==2019.3
sqlparse==0.3.1
